+++ src/allotropy/allotrope/converter.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 
 from collections.abc import Mapping
-from dataclasses import fields, is_dataclass, make_dataclass
+from dataclasses import asdict, fields, is_dataclass, make_dataclass
 from types import UnionType
-from typing import Any, Callable, cast, get_args, get_origin, Union
+from typing import Any, Callable, cast, get_args, get_origin, TypeVar, Union
 
 from cattrs import Converter
 from cattrs.errors import ClassValidationError
@@ -88,6 +88,15 @@ PRIMITIVE_TYPES = (
     np.float64,
     np.int64,
 )
+
+ModelClass = TypeVar("ModelClass")
+
+
+def add_custom_information_document(
+    model: ModelClass, custom_info_doc: dict[str, Any]
+) -> ModelClass:
+    model.custom_information_document = structure_custom_information_document(custom_info_doc, "custom information document")  # type: ignore
+    return model
 
 
 def _convert_model_key_to_dict_key(key: str) -> str:
@@ -282,11 +291,16 @@ def register_unstructure_hooks(converter: Converter) -> None:
             if not is_dataclass(obj):
                 return converter.unstructure(obj)
 
-            return {
+            dataclass_dict = {
                 _convert_model_key_to_dict_key(k): v
                 for k, v in make_unstructure_fn(type(obj))(obj).items()
                 if not should_omit(k, v)
             }
+            if hasattr(obj, "custom_information_document"):
+                dataclass_dict["custom information document"] = asdict(
+                    obj.custom_information_document
+                )
+            return dataclass_dict
 
         # This custom unstructure function overrides the unstruct_hook when we should should_allow_empty_value_field.
         # We need to do this at this level because we need to know both the parent class and the field name at the