+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added utility to add a custom information document to an ASM model
+
 ### Fixed
 
 - Updated schema cleaner to handle utf-8 characters in unit schema urls